Move Read Journal related classes to its own package
