+++ src/main/scala/akka/persistence/jdbc/dao/bytea/readjournal/ByteArrayReadJournalDao.scala
@@ -14,13 +14,13 @@
  * limitations under the License.
  */
 
-package akka.persistence.jdbc.dao.bytea
+package akka.persistence.jdbc.dao.bytea.readjournal
 
 import akka.NotUsed
 import akka.persistence.PersistentRepr
 import akka.persistence.jdbc.config.ReadJournalConfig
 import akka.persistence.jdbc.dao.ReadJournalDao
-import akka.persistence.jdbc.dao.bytea.ReadJournalTables.JournalRow
+import akka.persistence.jdbc.dao.bytea.readjournal.ReadJournalTables.JournalRow
 import akka.serialization.Serialization
 import akka.stream.Materializer
 import akka.stream.scaladsl.Source
